backend/app/database/db.py
from pymongo import MongoClient
from dotenv import load_dotenv
import os

load_dotenv()

_client = None

def get_db():
    global _client
    if _client is None:
        _client = MongoClient(os.getenv("MONGO_URI", "mongodb://localhost:27017/fincode_db"))
    return _client[os.getenv("MONGO_DB_NAME", "fincode_db")]
